build: modernize cmake pipeline, downgrade to c++20, and fix build failures
* default to ninja with fallback to make if unavailable
* build sdl3 statically (tray icon temporarily disabled due to crash)
* move resource zipping from shell script to cmake
* replace #embed with cmakerc for asset embedding
* inline shader source code directly instead of using #embed
* lower c++ standard requirement from c++26 to c++20
* update some icon sprites
* update dependency lists in readme
* fix compilation errors and warnings

+++ src/common/color.hpp
@@ -52,8 +52,8 @@ struct rgba {
     auto operator<=>(const rgba& rhs) const = default;
 
     rgba operator*(double n) const {
-      return rgba{std::clamp<i32>((i32)r * n, 0, 255), std::clamp<i32>((i32)g * n, 0, 255),
-                  std::clamp<i32>((i32)b * n, 0, 255), std::clamp<i32>((i32)a * n, 0, 255)};
+      return rgba{(u8)std::clamp<i32>((i32)r * n, 0, 255), (u8)std::clamp<i32>((i32)g * n, 0, 255),
+                  (u8)std::clamp<i32>((i32)b * n, 0, 255), (u8)std::clamp<i32>((i32)a * n, 0, 255)};
     }
 };
